+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* New user option `company-occurrence-weight-function`, allowing to tweak the
+  behavior of the transformer `company-sort-by-occurrence`.
 * Setting `company-idle-delay` to `t` is deprecated. Use the value 0 instead.
 
 ## 2014-07-01 (0.8.1)